requirements.txt
streamlit>=1.28.0
python-chess>=1.9.0
numpy>=1.24.0
matplotlib>=3.7.0
networkx>=3.0
pandas>=1.5.0
qiskit>=0.44.0
qiskit-aer>=0.12.0 